% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badparams.R
\docType{data}
\name{badparams}
\alias{badparams}
\title{Parameter names to be ignored when QC'ing SampleMaster data}
\format{
A character vector of parameter names
}
\source{
Internal package dataset
}
\usage{
badparams
}
\description{
This dataset contains parameter names that should be skipped during quality
control of SampleMaster laboratory data.
}
\keyword{internal}
