% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplemast_format.R
\name{calculate_discharge}
\alias{calculate_discharge}
\title{Calculate Stream Discharge from Cross-Sectional Measurements}
\usage{
calculate_discharge(
  data,
  location = "Location",
  date = "Date",
  branch = NULL,
  distance = "Distance from 0 (ft)",
  depth = "Depth (ft)",
  velocity = "Velocity (ft/s)",
  depth_unit = "ft",
  distance_unit = "ft",
  velocity_unit = "ft/s"
)
}
\arguments{
\item{data}{A data frame or tibble containing the required cross-sectional measurements.}

\item{location}{Name of the column identifying the sampling location. Default is `"Location"`.}

\item{date}{Name of the column identifying the date. Default is `"Date"`.}

\item{branch}{(Optional) Name of a numeric column identifying branches (may include any number of stream channels). Default is `NULL`.}

\item{distance}{(Required) Name of the column with distance from the stream's edge (in feet). Default is `"Distance from 0 (ft)"`.}

\item{depth}{(Required) Name of the column with measured depths (in feet). Default is `"Depth (ft)"`.}

\item{velocity}{(Required)Name of the column with measured water velocities (in ft/s). Default is `"Velocity (ft/s)"`.}

\item{depth_unit}{Character string identifying units used to measure depth . May be "ft" or "m".}

\item{distance_unit}{Character string identifying units used to measure segment width. May be "ft" or "m".}

\item{velocity_unit}{Character string identifying units used to measure velocity. May be "ft/s" or m/s"}
}
\value{
A tibble summarizing total discharge (in cubic feet per second) for each location and date.
}
\description{
Computes total stream discharge (flow) at each sampling location and date
using the midsection method (average depth of two adjacent locations), optionally
accounting for multiple branches.
}
\details{
This function applies the **midsection method** to estimate discharge using
measurements of depth and velocity across a stream transect. For each
segment between two distance points, the segment area is estimated as the
product of the segment width and the average depth. Segment discharge is
then computed as the product of area and mean velocity. Discharge is summed
across segments, and optionally, across branches.
}
\examples{
# Calculate discharge from example dataset
data(fieldvelocity)
discharge_results <- calculate_discharge(fieldvelocity,
                                         location = "Location",
                                         date = "Date",
                                         branch = "branch",
                                         distance = "Distance from 0 (ft)",
                                         depth = "Depth (ft)",
                                         velocity = "Velocity (ft/s)",
                                         depth_unit = "ft",
                                         distance_unit = "ft",
                                         velocity_unit = "ft/s")

# View a few rows
head(discharge_results)

}
