% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_patterns.R
\docType{data}
\name{dup_patterns}
\alias{dup_patterns}
\title{Possible entries to indicate field duplicate samples}
\format{
A character vector of possible field dup entries.
}
\source{
Internal package dataset
}
\usage{
dup_patterns
}
\description{
Possible entries to indicate field duplicate samples
}
\keyword{internal}
