% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaned_test_functions.R
\name{clean_sample_data}
\alias{clean_sample_data}
\title{Perform QC checks, and updates, on raw output from SampleMaster}
\usage{
clean_sample_data(return_QC_meta = TRUE)
}
\arguments{
\item{return_QC_meta}{Logical. If **TRUE**, saves a QC dataframe to the location of the
selected datafile with the name `filename_QC.xlsx`.}
}
\value{
A cleaned data frame with standardized column names and a `Warning` column listing any issues.
}
\description{
Performs standard cleaning and validation on sample data:
* Renames OrderDetails columns to standardized names.
* Removes rows with invalid `Param` values (from `badparams`) if present.
* Replaces field duplicate site names with their `Location`.
* Adds cumulative warnings for:
   - Remaining duplicate site indicators (e.g., "dup").
   - Missing or blank `Site` values.
}
\examples{
test_df <- data.frame(
  OrderDetails_User1 = 20,
  OrderDetails_User2 = "note",
  Site = c("Field Dup", "Station A", NA, ""),
  Location = c("Loc1", NA, "Loc3", "Loc4"),
  Param = c("NO2", "TP", "SRP", "PtCoN")
)

clean_sample_data(test_df)

}
